requirements.txt
numpy==1.21.5
streamlit==0.77.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
json5==0.9.6